[Wizard Boss] -Added Full functionality to Wizard -Minor Player Stats updateds ints to floats

+++ Assets/Scripts/AI/NPStats.cs
@@ -23,7 +23,7 @@ public class NPStats : UnitStats
         
     }
 
-    public override void ReceiveDamage(int damage)
+    public override void ReceiveDamage(float damage)
     {
         currentHealth -= damage;
         //StartCoroutine("turnON", 1);
